feat(ui): update translate file and move to component
