+++ src/ui/components/WalletConnectRequest/WalletConnectRequest.scss
