+++ src/ui/components/WalletConnectRequest/index.ts
